playing with node test runner
